migration/requirements.txt
psycopg2-binary>=2.9.6
python-tds>=1.11.0
#pytds>=1.12.0 
sqlglot>=10.6.2
python-dotenv>=0.19.2
tqdm>=4.66.1               # Added for progress bars
prometheus-client>=0.17.0  # Added for metrics
python_tds>=1.16.0
tenacity>=8.2.3
setuptools>=75.8.2
docopt>=0.6.2 
greenlet>=3.1.1 
migrator>=0.0.8 
sqlalchemy>=2.0.38
typing-extensions>=4.12.2
